SERVER: MongoDB integration, User collection - add, fetch

+++ server/src/application/useCases/user/verifyOtpAndCreateUser.ts
@@ -1,10 +1,12 @@
 import OtpRepository from '../../../domain/repositories/OtpRepository.js';
+import UserAuthRepository from '../../../domain/repositories/UserAuthRepository.js';
+import UserAuth from '../../../domain/entities/UserAuth.js';
 import UserRepository from '../../../domain/repositories/UserRepository.js';
-import User from '../../../domain/entities/User.js';
 
 class VerifyOtpAndCreateUser {
   constructor(
     private otpRepository: OtpRepository,
+    private userAuthRepository: UserAuthRepository,
     private userRepository: UserRepository,
   ) {}
 
@@ -15,18 +17,32 @@ class VerifyOtpAndCreateUser {
     lastName: string,
     password: string,
   ) {
-    const otp = await this.otpRepository.findByEmail(email);
-    if (!otp || otp.code !== otpCode) {
-      throw new Error('Invalid OTP');
-    }
+    try {
+      const otp = await this.otpRepository.findByEmail(email);
+      const userExist = await this.userAuthRepository.findByEmail(email);
 
-    if (otp.expiresAt < new Date()) {
-      throw new Error('OTP expired');
-    }
+      if (userExist) {
+        throw new Error('User already exists');
+      }
 
-    const user = new User(firstName, lastName, email, password);
-    await this.userRepository.create(user);
-    await this.otpRepository.deleteByEmail(email);
+      if (!otp || otp.code !== otpCode) {
+        throw new Error('Invalid OTP');
+      }
+
+      if (otp.expiresAt < new Date()) {
+        throw new Error('OTP expired');
+      }
+
+      const user = new UserAuth(firstName, lastName, email, password);
+      await this.userAuthRepository.create(user);
+      await this.userRepository.create(user);
+      await this.otpRepository.deleteByEmail(email);
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } catch (error: any) {
+      console.log('Error in verifying otp and creating user', error);
+
+      throw new Error(error);
+    }
   }
 }
